requirements.txt
google-generativeai
edge-tts
requests
shop
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
moviepy==1.0.3
ffmpeg-python
python-dotenv
